htpc-cec/requirements.txt
cec==0.2.8
python-daemon==3.1.2
